# backend/postman/specs/openapi.yaml
openapi: 3.0.3
info:
  title: MSM — Sistema MIP de Productividad y Reconocimiento Comercial
  description: |
    API del sistema MSM (Sistema MIP) para medir, gestionar y reconocer el
    desempeño de la fuerza de ventas farmacéutica.

    **Autenticación**: Bearer JWT. Usar `POST /auth/login` con form-data
    (`username` + `password`) para obtener el `access_token`.

    **Roles**: ADMIN · PRESIDENCIA · DIR_COMERCIAL · GERENTE_PRODUCTIVIDAD ·
    GERENTE_DISTRITO · GERENTE_MARCA · REPRESENTANTE_MEDICO · CONSULTA
  version: 1.0.0
  contact:
    email: segura.moises@gmail.com
servers:
  - url: http://localhost:8000/api/v1
    description: Desarrollo local
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    TokenResponse:
      type: object
      properties:
        access_token:  { type: string }
        refresh_token: { type: string }
        token_type:    { type: string, example: bearer }
        expires_in:    { type: integer, example: 3600 }
    Msg:
      type: object
      properties:
        message: { type: string }
    Error:
      type: object
      properties:
        detail: { type: string }
...
    UsuarioResponse:
      type: object
      properties:
        id:              { type: integer }
        username:        { type: string }
        email:           { type: string, nullable: true }
        nombre_completo: { type: string }
        rol:             { type: string }
        pais_id:         { type: integer, nullable: true }
        rm_id:           { type: integer, nullable: true }
        activo:          { type: boolean }
...
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Autenticación]
      summary: Iniciar sesión
      description: Autentica con username/password (form-data). Retorna JWT access + refresh token.
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string, example: admin }
                password: { type: string, example: Admin1234! }
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
              example:
                access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                token_type: bearer
                expires_in: 3600
        '401':
          description: Credenciales incorrectas
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { detail: Credenciales incorrectas }
        '422':
          description: Validación — campos requeridos ausentes
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '423':
          description: Cuenta bloqueada (5 intentos fallidos)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { detail: Cuenta bloqueada temporalmente. Intente más tarde. }
  /auth/logout:
    post:
      tags: [Autenticación]
      summary: Cerrar sesión
      description: Invalida el refresh_token en la blacklist.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token: { type: string }
      responses:
        '200':
          description: Sesión cerrada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Msg' }
        '401':
          description: Token inválido o expirado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/refresh:
    post:
      tags: [Autenticación]
      summary: Renovar access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '200':
          description: Nuevo access_token emitido
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '401':
          description: Refresh token inválido, expirado o revocado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { detail: Refresh token revocado. Inicie sesión nuevamente. }
  /auth/me:
    get:
      tags: [Autenticación]
      summary: Datos del usuario actual
      responses:
        '200':
          description: Usuario autenticado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UsuarioResponse' }
        '401':
          description: No autenticado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /dashboard/reconocimiento:
    get:
      tags: [Dashboard]
      summary: Dashboard de Reconocimiento (elegibles, premiados, certificados)
      parameters:
        - name: pais_id
          in: query
          schema: { type: integer }
        - name: ciclo_id
          in: query
          schema: { type: integer }
      responses:
        '200': { description: KPIs de reconocimiento }
...
  /reconocimiento:
    get:
      tags: [Reconocimiento]
      summary: Reconocimientos otorgados
      parameters:
        - name: pais_id
          in: query
          schema: { type: integer }
        - name: ciclo_id
          in: query
          schema: { type: integer }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: size
          in: query
          schema: { type: integer, default: 50 }
      responses:
        '200': { description: Lista de reconocimientos }
    post:
      tags: [Reconocimiento]
      summary: Registrar nuevo reconocimiento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rm_id, premio_id, ciclo_id]
              properties:
                rm_id:     { type: integer }
                premio_id: { type: integer }
                ciclo_id:  { type: integer }
      responses:
        '201': { description: Reconocimiento registrado }
        '400': { description: RM no elegible }

  # ─── HEALTH ───────────────────────────────────────────────────────────────
...